renamed multi* -> twoBoth*
multiAnonymous -> twoBothAnonymous
MultiUnnamed -> twoBothUnnamed
