+++ lang/lang.test/src/test/java/mb/pie/lang/test/funcDef/params/twoBothAnonymous/twoBothAnonymousComponent.java
@@ -1,4 +1,4 @@
-package mb.pie.lang.test.funcDef.params.multiAnonymous;
+package mb.pie.lang.test.funcDef.params.twoBothAnonymous;
 
 import dagger.Component;
 import mb.pie.dagger.PieComponent;
@@ -8,6 +8,6 @@ import javax.inject.Singleton;
 
 @mb.pie.dagger.PieScope
 @Component(modules = {PieModule.class, PieTestModule.class}, dependencies = {mb.log.dagger.LoggerComponent.class, mb.resource.dagger.ResourceServiceComponent.class})
-public interface multiAnonymousComponent extends PieComponent {
-    main_multiAnonymous get();
+public interface twoBothAnonymousComponent extends PieComponent {
+    main_twoBothAnonymous get();
 }
