+++ lang/lang.test/src/test/java/mb/pie/lang/test/funcDef/params/twoBothAnonymous/twoBothAnonymousTest.java
@@ -1,14 +1,14 @@
-package mb.pie.lang.test.funcDef.params.multiAnonymous;
+package mb.pie.lang.test.funcDef.params.twoBothAnonymous;
 
 import org.junit.jupiter.api.Test;
 
 import static mb.pie.lang.test.util.SimpleChecker.assertTaskOutputEquals;
 
-class multiAnonymousTest {
+class twoBothAnonymousTest {
     @Test void test() throws Exception {
         assertTaskOutputEquals(
-            DaggermultiAnonymousComponent.class,
-            new main_multiAnonymous.Input(2, null),
+            DaggertwoBothAnonymousComponent.class,
+            new main_twoBothAnonymous.Input(2, null),
             false
         );
     }
