+++ lang/lang.test/src/test/java/mb/pie/lang/test/funcDef/params/twoBothUnnamed/twoBothUnnamed.pie
@@ -1,0 +1,3 @@
+module mb:pie:lang:test:funcDef:params:twoBothUnnamed
+
+func main_twoBothUnnamed(path?, string) -> string = "red"
