+++ lang/lang.test/src/test/java/mb/pie/lang/test/funcDef/params/twoBothUnnamed/twoBothUnnamedComponent.java
@@ -1,4 +1,4 @@
-package mb.pie.lang.test.funcDef.params.MultiUnnamed;
+package mb.pie.lang.test.funcDef.params.twoBothUnnamed;
 
 import dagger.Component;
 import mb.pie.dagger.PieComponent;
@@ -8,6 +8,6 @@ import javax.inject.Singleton;
 
 @mb.pie.dagger.PieScope
 @Component(modules = {PieModule.class, PieTestModule.class}, dependencies = {mb.log.dagger.LoggerComponent.class, mb.resource.dagger.ResourceServiceComponent.class})
-public interface MultiUnnamedComponent extends PieComponent {
-    main_MultiUnnamed get();
+public interface twoBothUnnamedComponent extends PieComponent {
+    main_twoBothUnnamed get();
 }
