+++ lang/lang.test/src/test/java/mb/pie/lang/test/funcDef/params/twoBothUnnamed/twoBothUnnamedTest.java
@@ -1,0 +1,15 @@
+package mb.pie.lang.test.funcDef.params.twoBothUnnamed;
+
+import org.junit.jupiter.api.Test;
+
+import static mb.pie.lang.test.util.SimpleChecker.assertTaskOutputEquals;
+
+class twoBothUnnamedTest {
+    @Test void test() throws Exception {
+        assertTaskOutputEquals(
+            DaggertwoBothUnnamedComponent.class,
+            new main_twoBothUnnamed.Input(null, "hi"),
+            "red"
+        );
+    }
+}
